+++ src/board/automatic.rs
@@ -42,6 +42,7 @@ impl Plugin for AutomaticState {
 					handle_new_options,
 					handle_plane_clicked,
 					handle_cell_clicked,
+					ToggleAction::sys_action_hotkeys,
 				)
 					.in_set(OnUpdate(ProgramState::Automatic)),
 			)
@@ -145,6 +146,23 @@ impl ToggleAction {
 				}
 			}
 		});
+	}
+
+	pub fn sys_action_hotkeys(keys: Res<Input<KeyCode>>, mut selected_action: ResMut<Self>) {
+		for key in ToggleAction::iter() {
+			if keys.just_pressed(KeyCode::from(key)) {
+				*selected_action = key;
+			}
+		}
+	}
+}
+
+impl From<ToggleAction> for KeyCode {
+	fn from(value: ToggleAction) -> Self {
+			match value {
+				ToggleAction::TargetCell => KeyCode::T,
+				ToggleAction::ToggleCellEnabled => KeyCode::E,
+			}
 	}
 }