+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.1.12:
+- Hotkeys for targets/disable
+- Button to temporarily hide helper markers
+
 ## v0.1.11:
 - Copies state over from auto to manual
 - manual state: reset button